Printf

+++ Printf/ft_putchar.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-size_t	ft_putchar(char c)
+int	ft_putchar(char c)
 {
 	write(1, &c, 1);
 	return (1);
